Remove pointless tests
