+++ presto/tests/unit/test_sample.py
@@ -16,7 +16,6 @@ from presto.find_torsions import (
     DEFAULT_TORSIONS_TO_EXCLUDE_SMARTS,
     DEFAULT_TORSIONS_TO_INCLUDE_SMARTS,
 )
-from presto.mlp import get_ml_omm_system
 from presto.outputs import OutputType
 from presto.sample import (
     _SAMPLING_FNS_REGISTRY,
@@ -46,19 +45,6 @@ from presto.settings import (
     MMMDMetadynamicsTorsionMinimisationSamplingSettings,
     MMMDSamplingSettings,
     PreComputedDatasetSettings,
-)
-
-# Check if NNPOps is available (required for EGRET-1 and MACE models)
-try:
-    import NNPOps  # noqa: F401
-
-    NNPOPS_AVAILABLE = True
-except ImportError:
-    NNPOPS_AVAILABLE = False
-
-requires_nnpops = pytest.mark.skipif(
-    not NNPOPS_AVAILABLE,
-    reason="NNPOps not available (required for EGRET-1 and MACE models)",
 )
 
 
@@ -256,214 +242,6 @@ class TestLoadPrecomputedDataset:
                 device=device,
                 settings=settings,
                 output_paths={},
-            )
-
-
-class TestGetMlOmmSystem:
-    """Tests for get_ml_omm_system function."""
-
-    @pytest.fixture(autouse=True)
-    def mock_get_mlp(self):
-        """Mock get_mlp to avoid loading real models and pass isinstance checks."""
-        from unittest.mock import MagicMock, patch
-
-        import openmm
-
-        with patch("presto.mlp.get_mlp") as mock:
-            mock_potential = MagicMock()
-
-            def create_mock_system(topology, **kwargs):
-                system = openmm.System()
-                for _ in range(topology.getNumAtoms()):
-                    system.addParticle(1.0)
-                return system
-
-            mock_potential.createSystem.side_effect = create_mock_system
-            mock.return_value = mock_potential
-            yield mock
-
-    @requires_nnpops
-    def test_neutral_molecule_with_egret(self):
-        """Test creating system for neutral molecule with EGRET-1."""
-        mol = Molecule.from_smiles("CCO")
-        mol.generate_conformers(n_conformers=1)
-
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="egret-1"),
-            torch.device("cpu"),
-        )
-
-        assert isinstance(system, openmm.System)
-        assert system.getNumParticles() == mol.n_atoms
-
-    def test_neutral_molecule_with_aceff(self):
-        """Test creating system for neutral molecule with ACEFF-2.0."""
-        mol = Molecule.from_smiles("C")
-        mol.generate_conformers(n_conformers=1)
-
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aceff-2.0"),
-            torch.device("cpu"),
-        )
-
-        assert isinstance(system, openmm.System)
-        assert system.getNumParticles() == mol.n_atoms
-
-    def test_charged_molecule_with_aceff(self):
-        """Test creating system for charged molecule with ACEFF-2.0."""
-        mol = Molecule.from_smiles("[NH4+]")
-        mol.generate_conformers(n_conformers=1)
-
-        # Should not raise
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aceff-2.0"),
-            torch.device("cpu"),
-        )
-
-        assert isinstance(system, openmm.System)
-        assert system.getNumParticles() == mol.n_atoms
-
-    def test_charged_molecule_with_aimnet2(self):
-        """Test creating system for charged molecule with AIMNet2."""
-        mol = Molecule.from_smiles("[Cl-]")
-        mol.generate_conformers(n_conformers=1)
-
-        # Should not raise
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aimnet2"),
-            torch.device("cpu"),
-        )
-
-        assert isinstance(system, openmm.System)
-        assert system.getNumParticles() == mol.n_atoms
-
-    @pytest.mark.parametrize(
-        "smiles",
-        ["C", "CCO", "c1ccccc1", "CC(C)C"],
-    )
-    def test_various_neutral_molecules(self, smiles):
-        """Test various neutral molecules work with ACEFF-2.0."""
-        mol = Molecule.from_smiles(smiles)
-        mol.generate_conformers(n_conformers=1)
-
-        # Test with ACEFF-2.0 (doesn't require NNPOps)
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aceff-2.0"),
-            torch.device("cpu"),
-        )
-        assert isinstance(system, openmm.System)
-
-    @pytest.mark.parametrize(
-        "charged_smiles",
-        ["[NH4+]", "[Cl-]", "[Na+]", "[Ca+2]"],
-    )
-    def test_various_charged_molecules_with_compatible_models(self, charged_smiles):
-        """Test various charged molecules work with compatible models."""
-        mol = Molecule.from_smiles(charged_smiles)
-        mol.generate_conformers(n_conformers=1)
-
-        # Should work with charge-supporting models
-        system1 = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aceff-2.0"),
-            torch.device("cpu"),
-        )
-        system2 = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential="aimnet2"),
-            torch.device("cpu"),
-        )
-
-        assert isinstance(system1, openmm.System)
-        assert isinstance(system2, openmm.System)
-
-    @pytest.mark.parametrize(
-        "charged_smiles,model_name",
-        [
-            ("[NH4+]", "egret-1"),
-            ("[Cl-]", "mace-off23-small"),
-            ("[Na+]", "mace-off23-medium"),
-            ("[Ca+2]", "mace-off23-large"),
-        ],
-    )
-    def test_various_charged_molecules_with_any_model(self, charged_smiles, model_name):
-        """Test charged molecules are no longer blocked by presto allowlists."""
-        mol = Molecule.from_smiles(charged_smiles)
-        mol.generate_conformers(n_conformers=1)
-
-        system = get_ml_omm_system(
-            mol,
-            MLPSettings(ml_potential=model_name),
-            torch.device("cpu"),
-        )
-        assert isinstance(system, openmm.System)
-
-    def test_passes_potential_and_system_kwargs(self, mock_get_mlp):
-        """Test constructor and createSystem kwargs are forwarded."""
-        mol = Molecule.from_smiles("CCO")
-        mol.generate_conformers(n_conformers=1)
-
-        get_ml_omm_system(
-            mol,
-            MLPSettings(
-                ml_potential="custom-model",
-                ml_potential_kwargs={"modelPath": "my.model"},
-                ml_system_kwargs={"precision": "single"},
-            ),
-            torch.device("cpu"),
-        )
-
-        mock_get_mlp.assert_called_once_with("custom-model", modelPath="my.model")
-        create_kwargs = mock_get_mlp.return_value.createSystem.call_args.kwargs
-        assert create_kwargs["precision"] == "single"
-        assert create_kwargs["device"] == "cpu"
-        assert "charge" in create_kwargs
-
-    def test_ase_ml_system_kwargs_path(self, mock_get_mlp):
-        """Test ASE calculator/info kwargs in ml_system_kwargs are passed through."""
-        mol = Molecule.from_smiles("CCO")
-        mol.generate_conformers(n_conformers=1)
-        calculator = object()
-
-        get_ml_omm_system(
-            mol,
-            MLPSettings(
-                ml_potential="ase",
-                ml_system_kwargs={
-                    "calculator": calculator,
-                    "info": {"foo": "bar", "charge": 1},
-                },
-            ),
-            torch.device("cpu"),
-        )
-
-        mock_get_mlp.assert_called_once_with("ase")
-        create_kwargs = mock_get_mlp.return_value.createSystem.call_args.kwargs
-        assert create_kwargs["calculator"] is calculator
-        assert create_kwargs["info"] == {"foo": "bar", "charge": 1}
-        assert "charge" not in create_kwargs
-        assert "device" not in create_kwargs
-
-    def test_warns_charged_molecule_with_ase(self):
-        """Test charge warning for ASE path."""
-        mol = Molecule.from_smiles("[NH4+]")
-        mol.generate_conformers(n_conformers=1)
-
-        with pytest.warns(
-            UserWarning, match="does not automatically pass molecular charge"
-        ):
-            get_ml_omm_system(
-                mol,
-                MLPSettings(
-                    ml_potential="ase",
-                    ml_system_kwargs={"calculator": object()},
-                ),
-                torch.device("cpu"),
             )
